Add new validation tests.
